Extract Store::{patch_jobs} and related into impl module.
Following on from 81f0ec7949f16d10372a5f49243cd797f3c150ce. This time
for patch operations, further shrinking store.rs.
